+++ yo/app/scripts/controllers/cart.controller.js
@@ -195,7 +195,6 @@
                     if(!out) out = [];
                     _.each(out, function(cartItem){
                         cartItem.getSize(timeout.promise);
-                        cartItem.getStatus(timeout.promise);
                     });
                     defered.resolve(out);
                 });